# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: Interactive AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /access/verify:
    post:
      operationId: verifySerialVAccess
      tags: [openai]
      summary: Verify the private channel access key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessVerifyInput"
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessStatus"
        "401":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /access/logout:
    post:
      operationId: logoutSerialVAccess
      tags: [openai]
      summary: Close the private channel
      responses:
        "204":
          description: Access closed
  /access/status:
    get:
      operationId: getSerialVAccessStatus
      tags: [openai]
      summary: Check whether the private channel is open
      responses:
        "200":
          description: Current access state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessStatus"
  /openai/chat:
    post:
      operationId: sendOpenaiChat
      tags: [openai]
      summary: Stream a response from Serial Designation V
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiChatInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
        "500":
          description: AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OpenaiChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
      required:
        - role
        - content
    OpenaiChatInput:
      type: object
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 40
          items:
            $ref: "#/components/schemas/OpenaiChatMessage"
      required:
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AccessVerifyInput:
      type: object
      properties:
        password:
          type: string
          minLength: 1
      required:
        - password
    AccessStatus:
      type: object
      properties:
        granted:
          type: boolean
      required:
        - granted